requirements.txt
setuptools<82
pyramid_chameleon
pyramid_tm
zope.sqlalchemy
geoalchemy2
psycopg2-binary
